Datatype and enums changes

+++ services/payment-service/src/main/java/com/sj/ecommerce/payment_service/dto/CreatePaymentRequest.java
@@ -1,5 +1,3 @@
 package com.sj.ecommerce.payment_service.dto;
 
-import java.math.BigDecimal;
-
-public record CreatePaymentRequest(Long orderId, Long userId, BigDecimal amount, String idempotencyKey) {}
+public record CreatePaymentRequest(Long orderId, Long userId, Double amount, String idempotencyKey) {}
